# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "GitWeb's backend API for managing Git repositories and browsing the local filesystem.\nThis API provides endpoints for repository management, branch/commit operations, file management, and remote sync.",
        "title": "Gittyd API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@gitty.local"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8083",
    "basePath": "/",
    "paths": {
        "/api/auth/devices": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all device records without including token hashes. Requires Bearer auth.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "List paired devices",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/auth.DeviceTokenRecord"
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/devices/{deviceId}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Revokes the device token. Returns 204 No Content on success.",
                "tags": [
                    "auth"
                ],
                "summary": "Revoke a device",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device ID",
                        "name": "deviceId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "deviceId required",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "device not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/auth/local/pair": {
            "post": {
                "description": "Issues a device token for localhost connections after validating the master password. This is a simplified flow for the web frontend.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Localhost-only pairing",
                "parameters": [
                    {
                        "description": "Local pair request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.LocalPairRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.PairExchangeResponse"
                        }
                    },
                    "400": {
                        "description": "missing masterPassword",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "forbidden",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/auth/pair/exchange": {
            "post": {
                "description": "Validates the session and master password, then issues a new device token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Exchange pairing session for token",
                "parameters": [
                    {
                        "description": "Pair exchange request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.PairExchangeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.PairExchangeResponse"
                        }
                    },
                    "400": {
                        "description": "invalid JSON or missing required fields",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/auth/pair/session": {
            "get": {
                "description": "Returns the current pairing session information for QR code scanning",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Get current pairing session",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.PairSessionResponse"
                        }
                    },
                    "410": {
                        "description": "session expired or invalid",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "503": {
                        "description": "no active pairing session",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/filesystem/browse": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Browse a directory on the filesystem",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "filesystem"
                ],
                "summary": "Browse directory",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Directory path to browse",
                        "name": "path",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.DirectoryListing"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/filesystem/roots": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "List allowed root paths/volumes",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "filesystem"
                ],
                "summary": "List volume roots",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "auth.DeviceTokenRecord": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "description": "Token creation time in RFC3339 format",
                    "type": "string"
                },
                "deviceId": {
                    "description": "Device unique identifier",
                    "type": "string"
                },
                "deviceName": {
                    "description": "Human-readable device name",
                    "type": "string"
                },
                "lastUsedAt": {
                    "description": "Last time the token was used for authentication in RFC3339 format",
                    "type": "string"
                },
                "revokedAt": {
                    "description": "Time when the token was revoked in RFC3339 format",
                    "type": "string"
                },
                "tokenHash": {
                    "description": "SHA-256 hash of the bearer token (hex-encoded)",
                    "type": "string"
                }
            }
        },
...
        "handlers.LocalPairRequest": {
            "type": "object",
            "properties": {
                "masterPassword": {
                    "type": "string"
                }
            }
        },
        "handlers.PairExchangeRequest": {
            "type": "object",
            "properties": {
                "deviceName": {
                    "type": "string"
                },
                "masterPassword": {
                    "type": "string"
                },
                "sessionId": {
                    "type": "string"
                }
            }
        },
        "handlers.PairExchangeResponse": {
            "type": "object",
            "properties": {
                "deviceId": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "handlers.PairSessionResponse": {
            "type": "object",
            "properties": {
                "expiresAt": {
                    "type": "string"
                },
                "sessionId": {
                    "type": "string"
                }
            }
        },
...
        "models.DirectoryEntry": {
            "type": "object",
            "properties": {
                "is_directory": {
                    "type": "boolean",
                    "example": true
                },
                "is_git_repo": {
                    "type": "boolean",
                    "example": true
                },
                "is_hidden": {
                    "type": "boolean",
                    "example": false
                },
                "mod_time": {
                    "type": "string",
                    "example": "2024-01-15T10:30:00Z"
                },
                "name": {
                    "type": "string",
                    "example": "my-project"
                },
                "path": {
                    "type": "string",
                    "example": "/home/user/projects/my-project"
                },
                "permissions": {
                    "type": "string",
                    "example": "drwxr-xr-x"
                },
                "size": {
                    "type": "integer",
                    "example": 4096
                }
            }
        },
        "models.DirectoryListing": {
            "type": "object",
            "properties": {
                "can_go_up": {
                    "type": "boolean",
                    "example": true
                },
                "current_path": {
                    "type": "string",
                    "example": "/home/user/projects"
                },
                "entries": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.DirectoryEntry"
                    }
                },
                "parent_path": {
                    "type": "string",
                    "example": "/home/user"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter \"Bearer \u003ctoken\u003e\" where \u003ctoken\u003e is obtained from the /api/auth/pair/exchange endpoint",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}